+++ CHANGES.md
@@ -2,6 +2,8 @@
 -------------------
 - Require enclosure and telescope for instrument query
 - Avoid trying to reduce files that don't have fits extension in filename
+- Added entrypoint to update instruments table in database
+- Removed defunct entrypoint to reduce night by site
 
 0.19.3 (2019-02-13)
 -------------------